**Ver1.4** Feb.27th.2019
**Ver1.4** Feb.27th.2019

* Highly improved calling for SVA.
* Now PALMER supports other reference based bam files besides GRCh37, GRCh38 and hg19.
* Time consuming: to run PALMER on chr1/GRCh37, calling would cost ~24 hours (LINE-1/GRCh37), ~28 hours (Alu) or ~4 hours (SVA), for 8gb running memory minimun.
* A fatal bug fixed.
* Optimized scripts and outputs.
* Minor bugs fixed.
