+++ index/Alu.regions.GRCh38
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:5e3ab7a6cf976cf498d02fc52b522a0625319b1165f5284dccdbfc79807b833d
+size 37890694
