+++ index/region.split.index.GRCh37
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:60bdf59579581482929436b0fdee1ffca9daa695ebb391caec51c97d63db8da5
-size 64849
+oid sha256:bb8fe811ac7b6571264d12985c1e857d60cfbe9eb9a1c5e6f4dc7c06d2033f01
+size 64848
