+++ index/region.split.index.GRCh38
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:924590613744b32a6730283017c2528127d14f8d2bae2bd8f8b927f9885c0543
-size 73914
+oid sha256:b0aceafd5b19db50931648d85a3548a77af168a36be299493eb0ff070fcc3a86
+size 73913
