+++ index/region.split.index.length.hg19
@@ -1,0 +1,25 @@
+249250621
+135534747
+135006516
+133851895
+115169878
+107349540
+102531392
+90354753
+81195210
+78077248
+59128983
+243199373
+63025520
+48129895
+51304566
+198022430
+191154276
+180915260
+171115067
+159138663
+146364022
+141213431
+16569
+155270560
+59373566
