--- lib/SVA_E.fasta
@@ -1,21 +1,0 @@
->SVA_E
-ctctccctctccctctccctctccctctccctctccctctccctgtcccctctttccacggtctccctct
-gatgccgagccaaagctggacggtactgctgccatctcggctcactgcaacctccctgcctgattctcct
-gcctcagcctgccgagtgcctgcgattgcaggcgcgcgccgccacgcctgactggttttcgttttttttt
-ggtggagacggggtttcgctgtgttggccgggctggtctccagctcctaaccgcgagtgatccgccagcc
-tcggcctcccgaggtgccgggattgcagacggagtctcgttcactcagtgctcaatggtgcccaggctgg
-agtgcagtggcgtgatctcggctcgctacaaccacctcccagccgcctgccttggcctcccaaagagccg
-gattgcagcctctgcccggccgccaccccgtctgggaagtgaggagcgtctctgcctggccgcccatcgt
-ctgggatgtgaggagcccctctgcccggccgcccagtctgggaagtgaggagcgcctctgcccggccgcc
-atcccgtctaggaagtgaggagcgtctctgcccggccgcccatcgtctgagatgtggggagcgcctctgc
-cccgccgccccgtctgggatgtgaggagcgcctctgcccggccagccgccccgtctgggaggtggggggg
-tcagccccccgcccggccagccgccccgtccgggaggaggtgggggggtcagccccccgcccggccagcc
-gccccgtccgggaggtgaggggcgcctctgcccggccgcccctactgggaagtgaggagcccctctgccc
-ggccaccaccccgtctgggaggtgtacccagcggctcattggggatgggccatgatgacaatggcggttt
-tgtggaatagaaaggcgggaagggtggggaaaaaattgagaaatcggatggttgccgggtctgtgtggat
-agaagtagacatgggagacttttcattttgttctgtactaagaaaaattcttctgccttgggatcctgtt
-gatctgtgaccttatccccaaccctgtgctctctgaaacatgtgctgtgtccactcagggttaaatggat
-taagggcggtgcaagatgtgctttgttaaacagatgcttgaaggcagcatgctcgttaagagtcatcacc
-actccctaatcttaagtacccagggacacaaacactgcggaaggccgcagggtcctctgcctaggaaaac
-cagagacctttgttcacttgtttatctgctgaccttccctccactattgtcctatgaccctgccaaatcc
-ccctctgcgagaaacacccaagaatgatcaataaaaaaaaaaaaaaaaaaaa
